Create README.md
This website uses the COVID-19 API to generate charts related to the pandemic.
This is an HTML webpage that displays COVID-19 data using charts and tables. The webpage includes several JavaScript libraries, including Bootstrap, Chart.js, and Highcharts, which are used to create the charts.

The webpage displays a navigation bar at the top, which includes a search bar for searching for COVID-19 data for a specific country. The main content of the page is divided into several sections, each displaying a chart or table with COVID-19 data.

The JavaScript code included in the HTML file fetches data from the COVID-19 API and uses it to populate the charts and tables on the page. There are three URLs used to fetch data: one for global COVID-19 data, and two for COVID-19 data specific to Ethiopia. The data is fetched using the fetch() method and is returned in JSON format.

The data is then used to populate the charts and tables. The charts are created using Chart.js and Highcharts. Each chart has its own configuration, including the chart type, title, and data source. The tables are created using HTML and JavaScript, with the data being inserted into the table rows and cells using JavaScript.

The code first initializes an empty array dailyWorldData, an array of API URLs apiUrls, and an array of chart objects charts.

Each chart object in the charts array specifies the API URL to fetch data from, the ID of the canvas element that the chart will be drawn in, the title of the chart, and the type of chart. The code then loops through each chart object and fetches data from the corresponding API URL using fetch(). The data is then used to create a chart using Chart.js, which is drawn on the canvas element specified in the chart object.

The code also defines a getRandomColor() function that returns a random hex color code.

Finally, the code defines a search() function that generates an HTML figure element with an ID of "container". This function is not currently used or called in the code.

    <link
      href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
      rel="stylesheet"
      integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
      crossorigin="anonymous"
    />
This imports the Bootstrap CSS stylesheet from a CDN (Content Delivery Network). Bootstrap is a popular front-end framework that provides pre-built styles and components for web development.



    <script
      src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
      integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
      crossorigin="anonymous"
    ></script>
This imports the Bootstrap JavaScript library from a CDN. The JavaScript in this library is used to add interactivity and dynamic behavior to the webpage.


    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
This imports the Chart.js library from a CDN. Chart.js is a JavaScript library for creating interactive charts and graphs.



    <script src="https://code.highcharts.com/highcharts.js"></script>
    <script src="https://code.highcharts.com/modules/exporting.js"></script>
    <script src="https://code.highcharts.com/modules/export-data.js"></script>
    <script src="https://code.highcharts.com/modules/accessibility.js"></script>
These four script tags import the Highcharts library from a CDN. Highcharts is another JavaScript library for creating charts and graphs. 





Overall, this code is an example of a simple webpage that uses JavaScript libraries to display data in a visually appealing way.